Complete frontend debt ratchet batch 3
