+++ scripts/debt-ratchet-baseline.json
@@ -4,20 +4,11 @@
   "counts": {
     "any_casts": 0,
     "todo_comments": 0,
-    "oversized_files": 8,
+    "oversized_files": 0,
     "legacy_wordpress_runtime": 0,
     "bootstrap_imports": 0,
     "redux_imports": 0,
     "pages_router_files": 0
   },
-  "oversized_file_allowlist": [
-    "components/drops/view/item/content/attachments/DropAttachmentDisplay.tsx",
-    "components/drops/view/item/content/media/SeizeVideoPlayer.tsx",
-    "components/manifold-minting/ManifoldMinting.tsx",
-    "components/meme-calendar/meme-calendar.helpers.tsx",
-    "components/memelab/MemeLabPage.tsx",
-    "components/nextGen/admin/NextGenAdmin.tsx",
-    "components/nft-transfer/TransferModal.tsx",
-    "components/rememes/RememePage.tsx"
-  ]
+  "oversized_file_allowlist": []
 }
